Finished migrating tests
