+++ DinoTests/Steps/ControllerSteps.cs
@@ -23,10 +23,28 @@ namespace DinoTests.Steps
             sc["controller"].As<Controller>().gameState = Controller.status.running;
         }
 
+        [Given(@"a controller that can create objects on frame update")]
+        public void GivenAControllerThatCanCreateObjectsOnFrameUpdate()
+        {
+            IO IOdummy = new IO();
+            sc.Add("Controller", new Controller(IOdummy));
+        }
+        
+
         [When(@"the next frame cycle happens")]
         public void WhenTheNextFrameCycleHappens()
         {
             sc["controller"].As<Controller>().FrameUpdate();
+        }
+
+        [When(@"frameUpdate has been called (.*) times")]
+        public void WhenFrameUpdateHasBeenCalledTimes(int p0)
+        {
+            sc.Add("numberOfFrames", p0);
+            for (int i = 0; i < p0; i++)
+            {
+                sc["Controller"].As<Controller>().FrameUpdate();
+            }
         }
 
         [Then(@"no collision is detected")]
@@ -52,5 +70,42 @@ namespace DinoTests.Steps
         {
             sc.ContainsKey("Exception").Should().BeTrue();
         }
+
+        // This feature tests are best tested from a code review
+        /*
+        [Then(@"there should be close to (.*)% of the number of frameUpdates small rocks in the obstacles list")]
+        public void ThenThereShouldBeCloseToOfTheNumberOfFrameUpdatesSmallRocksInTheObstaclesList(Decimal p0)
+        {
+            int numberOfSmallRocks = 0;
+            float targetNumber = (float)p0 / 100 * sc["numberOfFrames"].As<int>();
+            sc["Controller"].As<Controller>().Obstacles.ForEach(
+                delegate (IPiece Obstacle)
+                {
+                    if (Obstacle is SmallRock)
+                    {
+                        numberOfSmallRocks++;
+                    }
+                });
+            numberOfSmallRocks.Should().BeInRange((int)(targetNumber - targetNumber / 10), (int)(targetNumber + targetNumber / 10));
+        }
+
+        [Then(@"there should be close to (.*)% of the number of frameUpdates Large rocks in the obstacles list")]
+        public void ThenThereShouldBeCloseToOfTheNumberOfFrameUpdatesLargeRocksInTheObstaclesList(Decimal p0)
+        {
+            ScenarioContext.Current.Pending();
+        }
+
+        [Then(@"there should be close to (.*)% of the number of frameUpdates birds in the obstacles list")]
+        public void ThenThereShouldBeCloseToOfTheNumberOfFrameUpdatesBirdsInTheObstaclesList(Decimal p0)
+        {
+            ScenarioContext.Current.Pending();
+        }
+
+        [Then(@"close to (.*) of those birds should be positioned at y = (.*)")]
+        public void ThenCloseToOfThoseBirdsShouldBePositionedAtY(int p0, int p1)
+        {
+            ScenarioContext.Current.Pending();
+        }
+        */
     }
 }
