--- DinoTests/Steps/JumperStepDefinitions.cs
@@ -1,93 +1,0 @@
-using System;
-using System.Collections.Generic;
-using TechTalk.SpecFlow;
-using DinoClassLib;
-using FluentAssertions;
-namespace DinoTests.Steps
-{
-    [Binding]
-    public sealed class JumperStepDefinitions
-    {
-        private int x, y;
-        // For additional details on SpecFlow step definitions see https://go.specflow.org/doc-stepdef
-
-        private readonly ScenarioContext sc;
-
-        public JumperStepDefinitions(ScenarioContext scenarioContext)
-        {
-            sc = scenarioContext;
-        }
-
-
-
-
-        
-
-
-
-
-
-
-
-
-
-
-
-        //big rock tests
-
-
-
-        
-
-
-
-
-
-
-
-
-
-        //bird tests start
-
-
-
-
-
-
-
-
-        //player tests
-
-
-
-
-
-
-
-
-
-
-        //frame update tests
-
-
-       
-
-        
-
-
-
-
-
-
-
-
-
-
-
-
- 
-
-
-
-    }
-}
